+++ src/WindowMaker.h
@@ -342,6 +342,7 @@ extern struct WPreferences {
 	char no_window_over_clip;
 	char minimize_hides_application;
 	char appicon_toggles_hide;
+	char touchpad_gestures;
 	WCoord window_place_origin;        /* Offset for windows placed on screen */
 
 	char constrain_window_size;        /* don't let windows get bigger than screen */